[L05***] implement reply operations

+++ L05/Controllers/ReplyController.cs
@@ -7,14 +7,114 @@ namespace L05.Controllers;
 
 public class ReplyController(AppDbContext db) : Controller
 {
-  public async Task<IActionResult> CreateAsync([FromBody] ReplyCreateDto dto)
+  [HttpGet]
+  public IActionResult Index([FromRoute] Guid id, [FromQuery] string? search)
   {
-    if (dto.Author == null || dto.Content == null)
+    string? nickname = HttpContext.Session.GetString("Nickname");
+    if (nickname == null)
     {
-      return BadRequest("Author and Content cannot be empty");
+      return RedirectToAction("Index", "Home");
     }
 
-    using AppDbContext db = new();
+    Models.Thread? thread = db.Threads
+      .Include(t => t.Replies)
+      .FirstOrDefault(t => t.Id == id);
+
+    if (thread == null) { return NotFound(); }
+
+    return View(new RepliesPageViewModel(
+      thread.Replies.ToList()
+        .Where(r =>
+          search == null || search == "" ||
+          r.Content.Contains(
+            search,
+            StringComparison.CurrentCultureIgnoreCase) ||
+          r.Author.Contains(
+            search,
+            StringComparison.CurrentCultureIgnoreCase))
+        .ToArray(),
+      id,
+      nickname,
+      search ?? ""));
+  }
+
+  [HttpGet]
+  public IActionResult Add([FromRoute] Guid id)
+  {
+    string? nickname = HttpContext.Session.GetString("Nickname");
+    if (nickname == null)
+    {
+      return RedirectToAction("Index", "Home");
+    }
+
+    ViewData["Title"] = "Add Reply";
+    return View(new ReplyCreateDto(
+      id,
+      "")
+    );
+  }
+
+  [HttpGet]
+  public IActionResult Edit([FromRoute] Guid id)
+  {
+    string? nickname = HttpContext.Session.GetString("Nickname");
+    if (nickname == null)
+    {
+      return RedirectToAction("Index", "Home");
+    }
+
+    Reply? reply = db.Replies
+      .FirstOrDefault(r => r.Id == id);
+
+    if (reply == null) { return NotFound(); }
+    if (nickname != reply.Author)
+    {
+      return RedirectToAction("Index", "Home");
+    }
+
+    ViewData["Title"] = "Edit Reply";
+    return View(new ReplyUpdateDto(
+      reply.Id,
+      reply.Content)
+    );
+  }
+
+  [HttpGet]
+  public IActionResult Delete([FromRoute] Guid id)
+  {
+    string? nickname = HttpContext.Session.GetString("Nickname");
+    if (nickname == null)
+    {
+      return RedirectToAction("Index", "Home");
+    }
+    Reply? reply = db.Replies
+      .FirstOrDefault(t => t.Id == id);
+
+    if (reply == null) { return NotFound(); }
+    if (nickname != reply.Author)
+    {
+      return RedirectToAction("Index", "Home");
+    }
+
+    ViewData["Title"] = "Delete Thread";
+    return View(new ReplyUpdateDto(
+      reply.Id,
+      reply.Content)
+    );
+  }
+
+  [HttpPost]
+  public async Task<IActionResult> CreateAsync([FromForm] ReplyCreateDto dto)
+  {
+    string? nickname = HttpContext.Session.GetString("Nickname");
+    if (nickname == null)
+    {
+      return RedirectToAction("Index", "Home");
+    }
+    if (dto.Content == null)
+    {
+      return BadRequest("Content cannot be empty");
+    }
 
     Models.Thread? thread = await db.Threads.FirstOrDefaultAsync(t => t.Id == dto.ThreadId);
     if (thread == null)
@@ -24,7 +124,7 @@ public class ReplyController(AppDbContext db) : Controller
 
     Reply reply = new()
     {
-      Author = dto.Author,
+      Author = nickname,
       Content = dto.Content,
       CreatedAt = DateTime.Now
     };
@@ -32,47 +132,64 @@ public class ReplyController(AppDbContext db) : Controller
 
     await db.SaveChangesAsync();
 
-    return Ok(reply);
+    return RedirectToAction(
+      "Index", "Reply",
+      new { id = dto.ThreadId });
   }
 
-  public async Task<IActionResult> GetAllByThreadIdAsync([FromRoute] Guid id)
-  {
-    Models.Thread? thread = await db.Threads
-      .Include(t => t.Replies)
-      .FirstOrDefaultAsync(t => t.Id == id);
-
-    if (thread == null) { return NotFound(); }
-
-    return Ok(thread.Replies.ToArray());
-  }
-
+  [HttpPost]
   public async Task<IActionResult> UpdateAsync(
-    [FromRoute] Guid id, [FromBody] ReplyUpdateDto dto)
+    [FromRoute] Guid id, [FromForm] ReplyUpdateDto dto)
   {
+    string? nickname = HttpContext.Session.GetString("Nickname");
+    if (nickname == null)
+    {
+      return RedirectToAction("Index", "Home");
+    }
+
     Reply? reply = await db.Replies
       .FirstOrDefaultAsync(r => r.Id == id);
 
     if (reply == null) { return NotFound(); }
+    if (nickname != reply.Author)
+    {
+      return RedirectToAction("Index", "Home");
+    }
 
     reply.Content = dto.Content;
     reply.UpdatedAt = DateTime.Now;
 
     await db.SaveChangesAsync();
 
-    return Ok(reply);
+    return RedirectToAction(
+      "Index", "Reply",
+      new { id = reply.ThreadId });
   }
 
+  [HttpPost]
   public async Task<IActionResult> DeleteAsync([FromRoute] Guid id)
   {
+    string? nickname = HttpContext.Session.GetString("Nickname");
+    if (nickname == null)
+    {
+      return RedirectToAction("Index", "Home");
+    }
+
     Reply? reply = await db.Replies
       .Include(r => r.Thread)
       .FirstOrDefaultAsync(r => r.Id == id);
 
     if (reply == null) { return NotFound(); }
+    if (nickname != reply.Author)
+    {
+      return RedirectToAction("Index", "Home");
+    }
 
     db.Replies.Remove(reply);
     await db.SaveChangesAsync();
 
-    return NoContent();
+    return RedirectToAction(
+      "Index", "Reply",
+      new { id = reply.ThreadId });
   }
 }
